Revert "ref(api): split api models into a class per file"
This reverts commit 004767d3327396af445b307cfb8de5deaccf98d0.
