+++ rootfs/api/tests/test_build.py
@@ -17,7 +17,7 @@ from api.models import Build
 from . import mock_status_ok
 
 
-@mock.patch('api.models.release.publish_release', lambda *args: None)
+@mock.patch('api.models.publish_release', lambda *args: None)
 class BuildTest(TransactionTestCase):
 
     """Tests build notification from build system"""
